Move Freetype to arte library.
