+++ src/lib/arte/arte/CMakeFinds.cmake.in
@@ -1,2 +1,4 @@
 @find_package@(Graphics CONFIG @REQUIRED@ COMPONENTS handy platform)
 @find_package@(Math CONFIG @REQUIRED@ COMPONENTS math)
+
+@find_package@(Freetype @REQUIRED@)
